\name{hypoData_mt}
\docType{data}
\alias{hypoData_mt}
\title{A simulation dataset for comparing two-group tag count
 data, focusing on RNA-seq}
\description{
A simulation dataset, consisting of 10,000 rows (or genes) and
 9 columns (or independent biological samples).
}
\usage{data(hypoData_mt)}
\format{hypoData is a matrix of dimension 10,000 times 9.}
\details{
This package typically start the differential expression analysis with
 a count table matrix such as hypoData_mt where each row indicates the
 gene (or transcript), each column indicates the sample (or library),
 and each cell indicates the number of counts to the gene in the sample.
 The first three columns are produced from biological replicates of,
 for example, Group 1, the next three columns are from Group2 and the
 remaining columns are from Group 3;
 i.e., {G1_rep1, G1_rep2, G1_rep3} vs. {G2_rep1, G2_rep2, G2_rep3} vs.
 {G3_rep1, G3_rep2, G3_rep3}.
 This data is generated by the \code{\link{generateSimulationData}}
 function with default parameter settings. The first 2,000 genes are
 differentially expressed among the three groups. Of these, the first 1,200
 genes are expressed at a higher level only in Group 1 (G1), the next
 600 genes are expressed at a higher level only in G2 and the remaining
 200 genes are expressed at a higher level only in G3. Accordingly, the
 2,001-10,000th genes are not differentially expressed (non-DEGs).
 The levels of differential expression (DE) are four-fold in both groups. 
} 
\examples{
\dontrun{
# The 'hypoData_mt' is generated by following commands.
tcc <- generateSimulationData(Ngene = 10000, PDEG = 0.2,
                              DEG.assign = c(0.9, 0.1),
                              DEG.model = "uniform",
                              DEG.foldchange = list(4, 4),
                              replicates = c(3, 3, 3))
hypoData_mt <- tcc$count
}
}
\keyword{datasets}
