\name{hypoData_ts}
\docType{data}
\alias{hypoData_ts}
\title{A sample microarray data for detecting tissue-specific patterns.}
\description{
A sample microarray data consists of eight rows (genes) and
ten columns (tissue).
}
\usage{data(hypoData_ts)}
\format{\code{hypoData_ts} is a matrix of dimension eight times ten.}
\examples{
data(hypoData_ts)
}
\keyword{datasets}
