\name{NBsample}
\alias{NBsample}
\title{Sampling from negative biomial distribution}
\description{
This methods allow sampling from Negative Binomial distribution with
specified proportion of differentially expressed genes having
specified level of differential expression in terms of fold change.
The proportion of upregulated are also specified.
The distribution of original expression levels are generated by resampling
real data of \emph{Arabidopsis} RNA-seq data from \code{\link{arab}}.
This function will be obsoleted. Use \link{simulateReadCounts}instead.
}
\usage{
NBsample(DEG_foldchange = 4, repA = 3, repB = 3, 
         Ngene = 3000, PDEG = 0.15, PA = 0.2)
}
\arguments{
    \item{DEG_foldchange}{Fold change value of differentially expressed genes}
    \item{repA}{Replicate number for sample A}
    \item{repB}{Replicate number for sample B}
    \item{Ngene}{Number of genes to produce}
    \item{PDEG}{Proportion of differentially expressed genes}
    \item{PA}{Proportion of upregulated genes in sample A among 
              differentially expressed genes (DEGs)}
}
\examples{
\dontrun{
sample <- NBsample()
}
}
