\name{nakai}
\docType{data}
\alias{nakai}
\title{DNA microarray data set}
\description{
This is a log2-transformed two-group microarray (Affymetrix GeneChip) data
consisting of 31,099 probesets. A total of eight samples were taken from the
rat liver: the first four samples are fed and the last fours are 24-hour
fasted. The original data can be obtained from NCBI Gene Expression Omnibus
(GEO) with GSE7623. This is a subset.
}
\usage{data(nakai)}
\format{\code{nakai} is a matrix of 31,099 rows (probesets) and 8 columns
(samples or tissues).}
\references{
Nakai Y, Hashida H., Kadota K, Minami M, Shimizu K, Matsumoto I, Kato H, Abe K., Up-regulation of genes related to the ubiquitin-proteasome system in the brown adipose tissue of 24-h-fasted rats. Bioscience Biotechnology and Biochemistry 2008, 72(1):139-148.
}
\examples{
data(nakai)
}
\keyword{datasets}
