\name{WAD}
\alias{WAD}
\title{Calculate WAD statistics for individual genes}
\description{
This methods allow sampling from Negative Binomial distribution with
specified proportion of differentially expressed genes having
specified level of differential expression in terms of fold change.
The proportion of upregulated are also specified.
The distribution of original expression levels are generated by resampling
real data of \emph{Arabidopsis} RNA-seq data from \code{\link{arab}}.
This function will be obsoleted. Use \link{simulateReadCounts}instead.
}
\usage{
WAD(data, group, k = 1)
}
\arguments{
    \item{data}{numeric matrix containing data.}
    \item{group}{numeric vector indicating the experimental group for each
                 sample (or library).}
    \item{k}{data equal to zero are placed by k.}
}
\examples{
data(hypoData)
group <- c(1, 1, 1, 2, 2, 2)

wad <- WAD(hypoData, group)
}
